# apifox-openapi.json
{"openapi":"3.1.0","info":{"title":"BizFormPlatform","version":"1.0.0"},"paths":{"/api/v1/forms/{form_type}/submit":{"post":{"tags":["Forms"],"summary":"Submit Merchant","description":"提交商户信息表单。","operationId":"submit_merchant_api_v1_forms__form_type__submit_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MerchantSubmit"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/forms/listing/submit":{"post":{"tags":["Forms"],"summary":"Submit Listing","description":"提交商户房源表单。","operationId":"submit_listing_api_v1_forms_listing_submit_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ListingSubmit"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/forms/product/submit":{"post":{"tags":["Forms"],"summary":"Submit Product","description":"提交商户商品表单。","operationId":"submit_product_api_v1_forms_product_submit_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductSubmit"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/forms/report/submit":{"post":{"tags":["Forms"],"summary":"Submit Report","description":"提交商户报表表单。","operationId":"submit_report_api_v1_forms_report_submit_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReportSubmit"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/forms/{form_type}":{"get":{"tags":["Forms"],"summary":"List Forms","operationId":"list_forms_api_v1_forms__form_type__get","parameters":[{"name":"form_type","in":"path","required":true,"schema":{"type":"string","title":"Form Type"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","default":1,"title":"Page"}},{"name":"page_size","in":"query","required":false,"schema":{"type":"integer","default":50,"title":"Page Size"}},{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"}},{"name":"keyword","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Keyword"}},{"name":"geohash_prefix","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Geohash Prefix"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/forms/{form_type}/{form_id}":{"get":{"tags":["Forms"],"summary":"Get Form","operationId":"get_form_api_v1_forms__form_type___form_id__get","parameters":[{"name":"form_type","in":"path","required":true,"schema":{"type":"string","title":"Form Type"}},{"name":"form_id","in":"path","required":true,"schema":{"type":"integer","title":"Form Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/forms/{form_type}/batch":{"post":{"tags":["Forms"],"summary":"Batch Process Forms","description":"Batch update form statuses with WebSocket progress.","operationId":"batch_process_forms_api_v1_forms__form_type__batch_post","parameters":[{"name":"form_type","in":"path","required":true,"schema":{"type":"string","title":"Form Type"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"array","items":{"type":"integer"},"title":"Form Ids"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/aggregation/run":{"post":{"tags":["Aggregation"],"summary":"Run Aggregation","description":"Step 1: Load submitted forms with geo → compute real geohash\nStep 2: Geohash rough bucketing (by prefix)\nStep 3: Haversine distance clustering within each bucket\nStep 4: Update batch_id and geohash in DB","operationId":"run_aggregation_api_v1_aggregation_run_post","parameters":[{"name":"distance_m","in":"query","required":false,"schema":{"type":"number","description":"聚合距离(米)","default":500,"title":"Distance M"},"description":"聚合距离(米)"},{"name":"precision","in":"query","required":false,"schema":{"type":"integer","description":"Geohash精度(字符数)","default":5,"title":"Precision"},"description":"Geohash精度(字符数)"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/aggregation/clusters":{"get":{"tags":["Aggregation"],"summary":"Get Clusters","description":"Get aggregation results grouped by cluster.","operationId":"get_clusters_api_v1_aggregation_clusters_get","parameters":[{"name":"batch_id","in":"query","required":false,"schema":{"type":"string","title":"Batch Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/knowledge/docs":{"post":{"tags":["Knowledge"],"summary":"Create Doc","operationId":"create_doc_api_v1_knowledge_docs_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DocCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["Knowledge"],"summary":"List Docs","operationId":"list_docs_api_v1_knowledge_docs_get","parameters":[{"name":"merchant_id","in":"query","required":false,"schema":{"type":"string","title":"Merchant Id"}},{"name":"category","in":"query","required":false,"schema":{"type":"string","title":"Category"}},{"name":"keyword","in":"query","required":false,"schema":{"type":"string","title":"Keyword"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"default":1,"title":"Page"}},{"name":"page_size","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":20,"title":"Page Size"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/knowledge/docs/{doc_id}":{"get":{"tags":["Knowledge"],"summary":"Get Doc","operationId":"get_doc_api_v1_knowledge_docs__doc_id__get","parameters":[{"name":"doc_id","in":"path","required":true,"schema":{"type":"integer","title":"Doc Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Knowledge"],"summary":"Update Doc","operationId":"update_doc_api_v1_knowledge_docs__doc_id__put","parameters":[{"name":"doc_id","in":"path","required":true,"schema":{"type":"integer","title":"Doc Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DocUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Knowledge"],"summary":"Delete Doc","operationId":"delete_doc_api_v1_knowledge_docs__doc_id__delete","parameters":[{"name":"doc_id","in":"path","required":true,"schema":{"type":"integer","title":"Doc Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/knowledge/docs/{doc_id}/versions":{"get":{"tags":["Knowledge"],"summary":"Get Doc Versions","operationId":"get_doc_versions_api_v1_knowledge_docs__doc_id__versions_get","parameters":[{"name":"doc_id","in":"path","required":true,"schema":{"type":"integer","title":"Doc Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/knowledge/docs/{doc_id}/rollback/{target_version}":{"post":{"tags":["Knowledge"],"summary":"Rollback Doc","description":"回退文档到指定版本——从快照恢复内容。","operationId":"rollback_doc_api_v1_knowledge_docs__doc_id__rollback__target_version__post","parameters":[{"name":"doc_id","in":"path","required":true,"schema":{"type":"integer","title":"Doc Id"}},{"name":"target_version","in":"path","required":true,"schema":{"type":"integer","title":"Target Version"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/analytics/submit-metrics":{"get":{"tags":["Analytics"],"summary":"Get Submit Metrics","description":"Get form submission success rate statistics.","operationId":"get_submit_metrics_api_v1_analytics_submit_metrics_get","parameters":[{"name":"form_type","in":"query","required":false,"schema":{"type":"string","title":"Form Type"}},{"name":"start","in":"query","required":false,"schema":{"type":"string","format":"date-time","title":"Start"}},{"name":"end","in":"query","required":false,"schema":{"type":"string","format":"date-time","title":"End"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/analytics/log":{"post":{"tags":["Analytics"],"summary":"Log Submission","description":"Record a form submission event.","operationId":"log_submission_api_v1_analytics_log_post","parameters":[{"name":"form_type","in":"query","required":true,"schema":{"type":"string","title":"Form Type"}},{"name":"merchant_id","in":"query","required":false,"schema":{"type":"string","title":"Merchant Id"}},{"name":"status","in":"query","required":false,"schema":{"type":"string","default":"start","title":"Status"}},{"name":"duration_ms","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Duration Ms"}},{"name":"error_code","in":"query","required":false,"schema":{"type":"string","title":"Error Code"}},{"name":"error_msg","in":"query","required":false,"schema":{"type":"string","title":"Error Msg"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"DocCreate":{"properties":{"merchant_id":{"type":"string","title":"Merchant Id"},"title":{"type":"string","title":"Title"},"content":{"type":"string","title":"Content"},"tags":{"items":{"type":"string"},"type":"array","title":"Tags","default":[]},"category":{"type":"string","title":"Category","default":"general"}},"type":"object","required":["merchant_id","title","content"],"title":"DocCreate"},"DocUpdate":{"properties":{"title":{"type":"string","title":"Title"},"content":{"type":"string","title":"Content"},"tags":{"items":{"type":"string"},"type":"array","title":"Tags"},"category":{"type":"string","title":"Category"},"status":{"type":"string","title":"Status"}},"type":"object","title":"DocUpdate"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ListingSubmit":{"properties":{"merchant_id":{"type":"integer","title":"Merchant Id","example":1},"title":{"type":"string","title":"Title","example":"新街口写字楼A座1201"},"area":{"type":"number","exclusiveMinimum":0.0,"title":"Area","example":180.0},"price":{"type":"number","exclusiveMinimum":0.0,"title":"Price","example":15000.0},"images":{"items":{"type":"string"},"type":"array","title":"Images","example":["img1.jpg"]}},"type":"object","required":["merchant_id","title","area","price"],"title":"ListingSubmit"},"MerchantSubmit":{"properties":{"name":{"type":"string","title":"Name","example":"星巴克新街口店"},"address":{"type":"string","title":"Address","example":"南京新街口正洪街18号"},"lat":{"type":"number","maximum":90.0,"minimum":-90.0,"title":"Lat","example":32.041},"lng":{"type":"number","maximum":180.0,"minimum":-180.0,"title":"Lng","example":118.784},"category":{"type":"string","title":"Category","default":"餐饮","example":"餐饮"},"contact_name":{"type":"string","title":"Contact Name","default":"","example":"张经理"},"contact_phone":{"type":"string","title":"Contact Phone","default":"","example":"13812345678"}},"type":"object","required":["name","address","lat","lng"],"title":"MerchantSubmit"},"ProductSubmit":{"properties":{"merchant_id":{"type":"integer","title":"Merchant Id","example":1},"name":{"type":"string","title":"Name","example":"拿铁咖啡"},"sku":{"type":"string","title":"Sku","example":"SKU-LATTE-001"},"category":{"type":"string","title":"Category","default":"饮品","example":"饮品"},"price":{"type":"number","exclusiveMinimum":0.0,"title":"Price","example":32.0},"stock":{"type":"integer","minimum":0.0,"title":"Stock","default":500,"example":500}},"type":"object","required":["merchant_id","name","sku","price"],"title":"ProductSubmit"},"ReportSubmit":{"properties":{"merchant_id":{"type":"integer","title":"Merchant Id","example":1},"report_type":{"type":"string","pattern":"^(daily|weekly|monthly)$","title":"Report Type","example":"monthly"},"period":{"type":"string","title":"Period","example":"2026-06"},"data":{"type":"object","title":"Data","example":{"orders":3200,"revenue":156000}}},"type":"object","required":["merchant_id","report_type","period"],"title":"ReportSubmit"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}